LayoutTests/fast/forms/date/calendar-picker-appearance-expected.txt

CONSOLE MESSAGE: line 17: Check if window.pagePopupController exists for a page-popup window.
CONSOLE MESSAGE: line 12: [object PagePopupController]
 
